Contain, center and enlarge MORE JELLIES button
The way that the flex grid is set up, we may want to consider changing two things. First, the number of jellies that get added per pagination to some number % 3 === 0 so we always have even flat rows. The second is the total number of jellies we have to follow the same rule. From initial load, 6 jellies are visible; possibly add the button there and keep adding 6 per click.